Move TacoResourceAssembler to resourceassembler package
